+++ storages/backends/overwrite.py
@@ -1,5 +1,13 @@
+import warnings
+
 from django.core.files.storage import FileSystemStorage
 from django.utils.deconstruct import deconstructible
+
+warnings.warn(
+    'OverwriteStorage is unmaintained and will be removed in the next django-storages version.'
+    'See https://github.com/jschneier/django-storages/issues/202',
+    PendingDeprecationWarning
+)
 
 
 @deconstructible