+++ storages/backends/symlinkorcopy.py
@@ -1,4 +1,5 @@
 import os
+import warnings
 
 from django.conf import settings
 from django.core.files.storage import FileSystemStorage
@@ -25,6 +26,11 @@ The rationale is that unmodified files will exist in their original location,
 e.g. /htdocs/example.com/image.jpg and modified files will be stored in
 a temporary directory, e.g. /tmp/image.jpg.
 """
+warnings.warn(
+    'SymlinkOrCopyStorage is unmaintained and will be removed in the next django-storages version.'
+    'See https://github.com/jschneier/django-storages/issues/202',
+    PendingDeprecationWarning
+)
 
 
 @deconstructible